make prompt editor, and fix the modal handlers in dream, aka refactor it and simplify it
