+++ src/components/shared/prompt-editor/index.ts
@@ -1,0 +1,2 @@
+export { PromptEditor, default } from "./prompt-editor";
+export type { PromptEditorProps } from "./prompt-editor";
